Scaffold and import font
To start this project I had to import some different fonts to make the website look edgy and cool. I also added the basics to the html file with the quickstart exclamation point and linked the css file style.css. The website is going to be very edgy but also very classy.