Add font.css
I understand how @font-face and @import work now: @font-face is located in the url that is used in @import. One COULD make font-faces for all of their custom fonts, but it would take up too much of the space recommended for the actual css styles. This file makes it easier to organize by having an entire css stylesheet specifically for fonts.